ML/hf_local/requirements.txt
transformers>=4.44.0
huggingface_hub>=0.24.0
torch>=2.2.0
Pillow>=10.0.0
requests>=2.31.0
librosa>=0.10.0
soundfile>=0.12.0
